# server/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is the API for PCBV. You can learn more on the GitHub page at\n\
    [https://github.com/LucasBaizer/pcbv](https://github.com/LucasBaizer/pcbv).\n"
  version: "1.0.0"
  title: "PCBV REST API"
  termsOfService: "http://swagger.io/terms"
  contact:
    email: "lucasallenbaizer@gmail.com"
  license:
    name: "MIT"
    url: "https://github.com/LucasBaizer/pcbv/tree/master/LICENSE.md"
basePath: "/api/v1"
tags:
- name: "circuit"
  description: "Work with circuits"
- name: "phrase"
schemes:
- "http"
- "https"
paths:
  /circuit:
    post:
      tags:
      - "circuit"
      summary: "Creates basic information and the images for a new circuit"
      operationId: "createCircuit"
      parameters:
      - in: "body"
        name: "body"
        description: "Data to create for the circuit"
        required: true
        schema:
          $ref: "#/definitions/Circuit"
      responses:
        200:
          description: "Circuit has been created"
          schema:
            $ref: "#/definitions/Circuit"
        400:
          description: "Invalid request data"
      x-swagger-router-controller: "Circuit"
  /circuit/{circuitId}:
    get:
      tags:
      - "circuit"
      summary: "Gets basic information and the image of a circuit"
      operationId: "getCircuit"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      responses:
        200:
          description: "Circuit is valid"
          schema:
            $ref: "#/definitions/Circuit"
        404:
          description: "Circuit does not exist"
      x-swagger-router-controller: "Circuit"
    delete:
      tags:
      - "circuit"
      summary: "Deletes a circuit"
      operationId: "deleteCircuit"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      responses:
        200:
          description: "Successfully deleted circuit"
        404:
          description: "Circuit does not exist"
      x-swagger-router-controller: "Circuit"
  /circuit/{circuitId}/subcircuit:
    get:
      tags:
      - "circuit"
      summary: "Gets basic information of all subcircuits of a circuit"
      operationId: "getSubCircuits"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "side"
        in: "query"
        description: "Side of the circuit the subcircuits are on"
        required: true
        type: "string"
        enum:
        - "front"
        - "back"
      responses:
        200:
          description: "Circuit is valid"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SubCircuit"
        404:
          description: "Circuit does not exist"
      x-swagger-router-controller: "Circuit"
    post:
      tags:
      - "circuit"
      summary: "Creates basic information and image of a particular subcircuit in\
        \ a circuit"
      operationId: "createSubCircuit"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Data to create for the subcircuit"
        required: true
        schema:
          $ref: "#/definitions/SubCircuit"
      - name: "side"
        in: "query"
        description: "Side of the circuit the subcircuit is on"
        required: true
        type: "string"
        enum:
        - "front"
        - "back"
      responses:
        200:
          description: "Successfully ceated"
          schema:
            $ref: "#/definitions/SubCircuit"
        400:
          description: "Invalid request data"
        404:
          description: "Circuit does not exist"
      x-swagger-router-controller: "Circuit"
  /circuit/{circuitId}/subcircuit/{subCircuitId}:
    get:
      tags:
      - "circuit"
      summary: "Gets image and basic information of a particular subcircuit of a circuit"
      operationId: "getSubCircuit"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "subCircuitId"
        in: "path"
        description: "ID of the subcircuit"
        required: true
        type: "integer"
      responses:
        200:
          description: "Circuit and subcircuit are valid"
          schema:
            $ref: "#/definitions/SubCircuit"
        404:
          description: "Either circuit or subcircuit do not exist"
      x-swagger-router-controller: "Circuit"
    delete:
      tags:
      - "circuit"
      summary: "Deletes a particular subcircuit of a circuit"
      operationId: "deleteSubCircuit"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "subCircuitId"
        in: "path"
        description: "ID of the subcircuit"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successfully deleted subcircuit"
        404:
          description: "Either circuit or subcircuit does not exist"
      x-swagger-router-controller: "Circuit"
  /circuit/{circuitId}/component:
    get:
      tags:
      - "circuit"
      summary: "Gets root components of a circuit"
      operationId: "getCircuitComponents"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "side"
        in: "query"
        description: "Side of the circuit the components are on"
        required: true
        type: "string"
        enum:
        - "front"
        - "back"
      responses:
        200:
          description: "Circuit is valid"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Component"
        404:
          description: "Circuit does not exist"
      x-swagger-router-controller: "Circuit"
    post:
      tags:
      - "circuit"
      summary: "Create a new root component for a circuit"
      operationId: "createComponent"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Data to create for the component"
        required: false
        schema:
          $ref: "#/definitions/Component"
      - name: "side"
        in: "query"
        description: "Side of the board the component is on"
        required: false
        type: "string"
        enum:
        - "front"
        - "back"
      responses:
        200:
          description: "Created component"
          schema:
            $ref: "#/definitions/Component"
        400:
          description: "Invalid request data"
        404:
          description: "Circuit does not exist"
      x-swagger-router-controller: "Circuit"
  /circuit/{circuitId}/component/{componentId}:
    post:
      tags:
      - "circuit"
      summary: "Updates the data for a particular component of a circuit"
      operationId: "updateCircuitComponent"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "componentId"
        in: "path"
        description: "ID of the component"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Data to update for the component"
        required: false
        schema:
          $ref: "#/definitions/Component"
      responses:
        200:
          description: "Successfully updated component"
          schema:
            $ref: "#/definitions/Component"
        400:
          description: "Invalid request data"
        404:
          description: "Either circuit of component do not exist"
      x-swagger-router-controller: "Circuit"
    delete:
      tags:
      - "circuit"
      summary: "Deletes a root component from the circuit"
      operationId: "deleteCircuitComponent"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "componentId"
        in: "path"
        description: "ID of the component"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successfully deleted component"
        404:
          description: "Either circuit or component does not exist"
      x-swagger-router-controller: "Circuit"
  /circuit/{circuitId}/subcircuit/{subCircuitId}/component:
    get:
      tags:
      - "circuit"
      summary: "Gets components of a particular subcircuit of a circuit"
      operationId: "getSubCircuitComponents"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "subCircuitId"
        in: "path"
        description: "ID of the subcircuit"
        required: true
        type: "integer"
      responses:
        200:
          description: "Circuit and subcircuit are valid"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Component"
        404:
          description: "Either circuit or subcircuit do not exist"
      x-swagger-router-controller: "Circuit"
    post:
      tags:
      - "circuit"
      summary: "Creates a component for a particular subcircuit of a circuit"
      operationId: "createSubCircuitComponent"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "subCircuitId"
        in: "path"
        description: "ID of the subcircuit"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Data to create for the component"
        required: true
        schema:
          $ref: "#/definitions/Component"
      responses:
        200:
          description: "Created component"
          schema:
            $ref: "#/definitions/Component"
        400:
          description: "Invalid request data"
        404:
          description: "Either circuit or subcircuit do not exist"
      x-swagger-router-controller: "Circuit"
  /circuit/{circuitId}/subcircuit/{subCircuitId}/component/{componentId}:
    post:
      tags:
      - "circuit"
      summary: "Updates the data for a particular component of a particular subcircuit\
        \ of a circuit"
      operationId: "updateSubCircuitComponent"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "subCircuitId"
        in: "path"
        description: "ID of the subcircuit"
        required: true
        type: "integer"
      - name: "componentId"
        in: "path"
        description: "ID of the component"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Data to update for the component"
        required: false
        schema:
          $ref: "#/definitions/Component"
      responses:
        200:
          description: "Successfully updated component"
          schema:
            $ref: "#/definitions/Component"
        400:
          description: "Invalid request data"
        404:
          description: "Either circuit, subcircuit, or component do not exist"
      x-swagger-router-controller: "Circuit"
    delete:
      tags:
      - "circuit"
      summary: "Deletes a component from a particular subcircuit of a circuit"
      operationId: "deleteSubCircuitComponent"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "subCircuitId"
        in: "path"
        description: "ID of the subcircuit"
        required: true
        type: "integer"
      - name: "componentId"
        in: "path"
        description: "ID of the component"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successfully deleted component"
        404:
          description: "Either circuit, subcircuit, or component do not exist"
      x-swagger-router-controller: "Circuit"
  /circuit/{circuitId}/category:
    get:
      tags:
      - "circuit"
      summary: "Gets categories of a circuit"
      operationId: "getCircuitCategories"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      responses:
        200:
          description: "Circuit is valid"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Category"
        404:
          description: "Circuit does not exist"
      x-swagger-router-controller: "Circuit"
    post:
      tags:
      - "circuit"
      summary: "Creates a new circuit category"
      operationId: "createCircuitCategory"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Data to create for the category"
        required: false
        schema:
          $ref: "#/definitions/Category"
      responses:
        200:
          description: "Successfully created circuit"
          schema:
            $ref: "#/definitions/Category"
        400:
          description: "Invalid request data"
        404:
          description: "Circuit does not exist"
      x-swagger-router-controller: "Circuit"
  /circuit/{circuitId}/category/{categoryId}:
    post:
      tags:
      - "circuit"
      summary: "Updates data of particular category of a circuit"
      operationId: "updateCircuitCategory"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "categoryId"
        in: "path"
        description: "ID of the category"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Data to update for the category"
        required: true
        schema:
          $ref: "#/definitions/Category"
      responses:
        200:
          description: "Successfully updated category"
          schema:
            $ref: "#/definitions/Category"
        400:
          description: "Invalid request data"
        404:
          description: "Either circuit or category does not exist"
      x-swagger-router-controller: "Circuit"
    delete:
      tags:
      - "circuit"
      summary: "Deletes a particular category of a circuit"
      operationId: "deleteCircuitCategory"
      parameters:
      - name: "circuitId"
        in: "path"
        description: "GUID of the circuit"
        required: true
        type: "string"
      - name: "categoryId"
        in: "path"
        description: "ID of the category"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successfully deleted category"
        404:
          description: "Either circuit or category does not exist"
      x-swagger-router-controller: "Circuit"
  /phrase:
    get:
      tags:
      - "phrase"
      summary: "Gets a PCBV random phrase"
      description: "Gets a random phrase, which has an acronym of P.C.B.V.\n"
      operationId: "getPcbvPhrase"
      parameters: []
      responses:
        200:
          description: "Successfully got the random phrase"
          schema:
            $ref: "#/definitions/PcbvPhrase"
      x-swagger-router-controller: "Phrase"
definitions:
  Circuit:
    type: "object"
    properties:
      circuitId:
        type: "string"
        example: "505b791a-b324-4880-8b76-aeca2e054dca"
      name:
        type: "string"
        example: "My PCB"
      imageFront:
        type: "string"
        example: "abcdef=="
      imageBack:
        type: "string"
        example: "abcdef=="
      imageType:
        type: "string"
        example: "png"
    example:
      imageFront: "abcdef=="
      name: "My PCB"
      imageBack: "abcdef=="
      imageType: "png"
      circuitId: "505b791a-b324-4880-8b76-aeca2e054dca"
  SubCircuit:
    type: "object"
    properties:
      subCircuitId:
        type: "integer"
        format: "int32"
        example: 1
      parentCircuitId:
        type: "string"
        example: "505b791a-b324-4880-8b76-aeca2e054dca"
      image:
        type: "string"
        example: "abcdef=="
      imageType:
        type: "string"
        example: "png"
      bounds:
        $ref: "#/definitions/ComponentBounds"
      side:
        type: "string"
        enum:
        - "front"
        - "back"
    example:
      image: "abcdef=="
      side: "front"
      parentCircuitId: "505b791a-b324-4880-8b76-aeca2e054dca"
      bounds:
        x: 500
        width: 200
        y: 1000
        height: 200
      subCircuitId: 1
      imageType: "png"
  Component:
    type: "object"
    properties:
      componentId:
        type: "integer"
        format: "int32"
        example: 1
      subCircuitId:
        type: "integer"
        format: "int32"
        example: 1
      cirucitId:
        type: "string"
        example: "505b791a-b324-4880-8b76-aeca2e054dca"
      bounds:
        $ref: "#/definitions/ComponentBounds"
      name:
        type: "string"
        example: "My Component"
      description:
        type: "string"
        example: "My Component is used do to x, y, and z"
      documentationUrl:
        type: "string"
        example: "http://example.com"
      category:
        $ref: "#/definitions/Category"
    example:
      documentationUrl: "http://example.com"
      componentId: 1
      cirucitId: "505b791a-b324-4880-8b76-aeca2e054dca"
      bounds:
        x: 500
        width: 200
        y: 1000
        height: 200
      name: "My Component"
      subCircuitId: 1
      description: "My Component is used do to x, y, and z"
      category:
        color: "FF00FF"
        name: "Memory"
        categoryId: 1
        circuitId: "505b791a-b324-4880-8b76-aeca2e054dca"
        tags:
        - "ddr"
        - "gb"
  ComponentBounds:
    type: "object"
    properties:
      x:
        type: "integer"
        example: 500
      y:
        type: "integer"
        example: 1000
      width:
        type: "integer"
        example: 200
      height:
        type: "integer"
        example: 200
    example:
      x: 500
      width: 200
      y: 1000
      height: 200
  Category:
    type: "object"
    properties:
      categoryId:
        type: "integer"
        format: "int32"
        example: 1
      circuitId:
        type: "string"
        example: "505b791a-b324-4880-8b76-aeca2e054dca"
      name:
        type: "string"
        example: "Memory"
      color:
        type: "string"
        example: "FF00FF"
      tags:
        type: "array"
        example:
        - "ddr"
        - "gb"
        items:
          type: "string"
    example:
      color: "FF00FF"
      name: "Memory"
      categoryId: 1
      circuitId: "505b791a-b324-4880-8b76-aeca2e054dca"
      tags:
      - "ddr"
      - "gb"
  PcbvPhrase:
    type: "object"
    properties:
      phrase:
        type: "string"
    example:
      phrase: "phrase"
externalDocs:
  description: "Find out more about PCBV"
  url: "https://github.com/LucasBaizer/pcbv"